Added solution of new task 59 New package "GroupTasks" was created and all class tasks was moved into this package
