+++ src/Dmitrii/GroupTasks/Task_80.java
@@ -1,4 +1,4 @@
-package Dmitrii;
+package Dmitrii.GroupTasks;
 
 import java.util.*;
